renamed `FamilyData` to `ClientData` && renamed state variables that reference 'family' into 'client'
